refactor (tests): Share EntanglerCore test funcs
Move common EntanglerCore functions to a shared file,
test/helpers/coretester.py. Funcs are same as before, just have them
in a shared location.
