+++ test/test_entangler_core.py
@@ -10,12 +10,12 @@ sys.path.append(os.path.join(os.path.dirname(__file__), "helpers"))
 
 
 from dynaconf import settings  # noqa: E402
-from migen import Module  # noqa: E402
 from migen import run_simulation  # noqa: E402
 from migen import Signal  # noqa: E402
 
 # pylint: disable=import-error
 from entangler.core import EntanglerCore  # noqa: E402
+from coretester import CoreTestHarness  # noqa: E402
 from gateware_utils import MockPhy  # noqa: E402 ./helpers/gateware_utils
 from gateware_utils import advance_clock  # noqa: E402
 from gateware_utils import wait_until  # noqa: E402
@@ -30,7 +30,7 @@ def int_to_bool_array(val: int, num_binary_digits: int) -> typing.Sequence[bool]
     return bool_arr
 
 
-class StandaloneHarness(Module):
+class StandaloneHarness(CoreTestHarness):
     """Test harness for the ``EntanglerCore``."""
 
     def __init__(self):
@@ -57,75 +57,6 @@ class StandaloneHarness(Module):
         )
 
         self.comb += self.counter.eq(self.core.msm.m)
-
-    def setup_core(self, cycle_length: int, timeout: int):
-        """Initialize the basic settings for the ``EntanglerCore``."""
-        msm = self.core.msm
-        yield msm.cycle_length_input.eq(cycle_length)
-        yield msm.timeout_input.eq(timeout)
-        yield msm.is_master.eq(1)
-        yield msm.standalone.eq(1)
-
-    def set_sequencer_outputs(
-        self, time_pairs: typing.Sequence[typing.Tuple[int, int]]
-    ) -> None:
-        """Set output TTL/GPIO timings."""
-        sequencers = self.core.sequencers
-        i = -1
-        for i, timing_pair in enumerate(time_pairs):
-            start, stop = timing_pair
-            yield sequencers[i].m_start.eq(start)
-            yield sequencers[i].m_stop.eq(stop)
-        for disable_ind in range(i + 1, len(sequencers)):
-            yield sequencers[disable_ind].m_start.eq(0)
-            yield sequencers[disable_ind].m_stop.eq(0)
-
-    def set_gating_times(
-        self, time_pairs: typing.Sequence[typing.Tuple[int, int]]
-    ) -> None:
-        """Set time windows when the input gaters will register input events."""
-        gaters = self.core.apd_gaters
-        i = -1
-        for i, timing_pair in enumerate(time_pairs):
-            start, stop = timing_pair
-            yield gaters[i].gate_start.eq(start)
-            yield gaters[i].gate_stop.eq(stop)
-        for disable_ind in range(i + 1, len(gaters)):
-            yield gaters[disable_ind].gate_start.eq(0)
-            yield gaters[disable_ind].gate_stop.eq(0)
-
-    def set_event_times(self, event_times: typing.Sequence[int]) -> None:
-        """Set the times when the mocked 'input' signals will occur."""
-        for i, time in enumerate(event_times):
-            yield getattr(self, "phy_apd{}".format(i)).t_event.eq(time)
-
-    def set_patterns(self, pattern_list: typing.Sequence[int]):
-        """Set the patterns that the ``EntanglerCore`` will try to match."""
-        patterns = self.core.heralder.patterns
-        enables = self.core.heralder.pattern_ens
-        for i, pattern in enumerate(pattern_list):
-            assert pattern < 2 ** len(patterns[i])
-            _LOGGER.debug("Setting pattern %i = %x", i, pattern)
-            yield patterns[i].eq(pattern)
-            # yield
-            # assert (yield patterns[i]) == pattern
-        # set enables. Convert # of patterns -> one-hot encoding
-        yield enables.eq((2 ** len(pattern_list)) - 1)
-
-        # Verify enable setting
-        # yield
-        # _LOGGER.debug(
-        #     "Enables val: %i, should be %i",
-        #     (yield enables),
-        #     ((2 ** len(pattern_list) - 1)),
-        # )
-        # assert (yield enables) == (2 ** len(pattern_list) - 1)
-
-    def start_entanglement_generator(self) -> None:
-        """Start the state machine that generates & checks for entanglement."""
-        yield self.core.msm.run_stb.eq(1)
-        yield
-        yield self.core.msm.run_stb.eq(0)
 
 
 def standalone_test(dut: StandaloneHarness):
